# RedisEncoderBenchmark