@Scottmitch 2016-05-04
@normanmaurer - this is to benchmark https://github.com/netty/netty/pull/5187