@Scottmitch 2016-05-04
current 4.1 branch
```
Benchmark                         (arraySize)  (pooledAllocator)  (voidPromise)   Mode  Cnt      Score       Error  Units
RedisEncoderBenchmark.writeArray           50               true           true  thrpt    5  63969.220 ±  6382.102  ops/s
RedisEncoderBenchmark.writeArray           50               true          false  thrpt    5  49814.431 ± 12195.522  ops/s
RedisEncoderBenchmark.writeArray           50              false           true  thrpt    5  33211.060 ±  3504.983  ops/s
RedisEncoderBenchmark.writeArray           50              false          false  thrpt    5  30743.100 ±   499.501  ops/s
RedisEncoderBenchmark.writeArray          200               true           true  thrpt    5  15591.464 ±  3394.825  ops/s
RedisEncoderBenchmark.writeArray          200               true          false  thrpt    5  13390.514 ±   379.690  ops/s
RedisEncoderBenchmark.writeArray          200              false           true  thrpt    5   8147.296 ±   952.575  ops/s
RedisEncoderBenchmark.writeArray          200              false          false  thrpt    5   7865.048 ±  1073.804  ops/s
RedisEncoderBenchmark.writeArray         1000               true           true  thrpt    5   2544.127 ±    51.005  ops/s
RedisEncoderBenchmark.writeArray         1000               true          false  thrpt    5   2198.244 ±   265.702  ops/s
RedisEncoderBenchmark.writeArray         1000              false           true  thrpt    5   1469.964 ±   258.873  ops/s
RedisEncoderBenchmark.writeArray         1000              false          false  thrpt    5   1712.845 ±    62.373  ops/s
```

pr #5187 
```
Benchmark                         (arraySize)  (pooledAllocator)  (voidPromise)   Mode  Cnt      Score       Error  Units
RedisEncoderBenchmark.writeArray           50               true           true  thrpt    5  68281.282 ± 11148.027  ops/s
RedisEncoderBenchmark.writeArray           50               true          false  thrpt    5  62231.308 ± 10257.768  ops/s
RedisEncoderBenchmark.writeArray           50              false           true  thrpt    5  32490.653 ±  3856.137  ops/s
RedisEncoderBenchmark.writeArray           50              false          false  thrpt    5  31663.050 ±   561.172  ops/s
RedisEncoderBenchmark.writeArray          200               true           true  thrpt    5  16371.227 ±  1670.922  ops/s
RedisEncoderBenchmark.writeArray          200               true          false  thrpt    5  14778.851 ±  1449.987  ops/s
RedisEncoderBenchmark.writeArray          200              false           true  thrpt    5   8524.627 ±  1167.727  ops/s
RedisEncoderBenchmark.writeArray          200              false          false  thrpt    5   8533.671 ±   523.321  ops/s
RedisEncoderBenchmark.writeArray         1000               true           true  thrpt    5   3096.848 ±   586.170  ops/s
RedisEncoderBenchmark.writeArray         1000               true          false  thrpt    5   3189.777 ±   303.031  ops/s
RedisEncoderBenchmark.writeArray         1000              false           true  thrpt    5   1720.446 ±    30.271  ops/s
RedisEncoderBenchmark.writeArray         1000              false          false  thrpt    5   1732.278 ±   115.518  ops/s
```

Improvement in this benchmark is as much as ~45%. These results likely also don't demonstrate the savings in GC due to allocating less promise objects.