@normanmaurer 2017-12-16
I removed the unused imports and merged it into 4.1 (5f0342ebe065f692038a902216dd9f5b02e0a2ef). Cant hurt.